The bonus level using portals is done
It works using the first collision as a teleport to another level with the original level being saved as the next level for when the coins are collected and the player collides with the portal again
Then, the portal is deactivated and it returns you to your original level.

It took a lot of Interconnecting between the level system and the collision system along with a lot of Boolean variables.